pyproject.toml
[project]
name = "flux-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "google-generativeai==0.8.3",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
    "uvicorn[standard]==0.30.6",
    "duckduckgo-search>=8.0.0",
    "openai>=2.21.0",
    "beautifulsoup4>=4.14.3",
    "streamlit>=1.54.0",
    "twilio>=9.10.2",
    "python-telegram-bot>=22.6",
]
